[babel-plugin] Rename unit test files
Remove 'stylex' prefix from test names
