+++ packages/babel-plugin/__tests__/metadata-test.js
