+++ packages/babel-plugin/__tests__/transform-define-vars-test.js
