+++ packages/babel-plugin/__tests__/transform-logical-properties-test.js
