+++ packages/babel-plugin/__tests__/transform-value-normalize-test.js
