+++ packages/babel-plugin/__tests__/transform-variable-removal-test.js
