+++ packages/babel-plugin/__tests__/validation-custom-properties-test.js
