+++ packages/babel-plugin/__tests__/validation-declarations-test.js
