+++ packages/babel-plugin/__tests__/validation-define-vars-test.js
